Extracted confirmation token handling logic from UserService
