+++ hotel-booking-app-domain/src/main/java/com/application/hotelbooking/repositories/UserRepository.java
@@ -12,4 +12,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
     Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
 }
