+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/services/UserEmailConfirmationService.java
@@ -1,0 +1,10 @@
+package com.application.hotelbooking.services;
+
+import org.springframework.transaction.annotation.Transactional;
+
+public interface UserEmailConfirmationService {
+    @Transactional
+    void confirmToken(String token);
+    @Transactional
+    void resendConfirmationToken(String email);
+}
