+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/services/repositoryservices/UserRepositoryService.java
@@ -8,4 +8,5 @@ public interface UserRepositoryService {
     Optional<UserModel> getUserByName(String username);
     Optional<UserModel> getUserByEmail(String email);
     UserModel save(UserModel userModel);
+    boolean emailExists(String email);
 }
