+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/services/repositoryservices/implementations/UserRepositoryServiceImpl.java
@@ -29,4 +29,8 @@ public class UserRepositoryServiceImpl implements UserRepositoryService {
     public UserModel save(UserModel userModel){
         return userTransformer.transformToUserModel(userRepository.save(userTransformer.transformToUser(userModel)));
     }
+
+    public boolean emailExists(String email) {
+        return userRepository.existsByEmail(email);
+    }
 }
